requirements.txt
python-dotenv
tk
requests
markdown2
tkhtmlview
transformers
torch
sentencepiece
multiprocessing